# dashboard/docs/api-spec.yaml
openapi: 3.0.3
info:
  title: Memory Observer Dashboard API
  description: |
    REST API for the Memory Observer Dashboard, providing access to workspaces, plans, agents, and events
    from the Project Memory MCP Server.
    
    ## Overview
    
    The Memory Observer Dashboard API allows you to:
    - View and manage MCP workspaces
    - Browse, create, and manage project plans
    - View and deploy agent templates
    - Subscribe to real-time MCP events
    - Search across all workspaces and plans
    
    ## Authentication
    
    Currently, the API runs locally and does not require authentication.
    Future versions may support authentication for remote access.
    
  version: 0.1.0
  contact:
    name: Project Memory MCP
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3001
    description: Local Development Server
...
paths:
  /api/workspaces:
    get:
      tags:
        - Workspaces
      summary: List all workspaces
      description: Returns all registered workspaces with their metadata
      responses:
        '200':
          description: List of workspaces
          content:
            application/json:
              schema:
                type: object
                properties:
                  workspaces:
                    type: array
                    items:
                      $ref: '#/components/schemas/Workspace'
                  total:
                    type: integer
                    example: 5
        '500':
          $ref: '#/components/responses/InternalError'
  /api/workspaces/{workspaceId}:
    get:
      tags:
        - Workspaces
      summary: Get workspace details
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Workspace details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WorkspaceDetails'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/workspaces/{workspaceId}/philosophy:
    get:
      tags:
        - Workspaces
      summary: Get project philosophy file
      description: Retrieves the project-philosophy.md content for a workspace
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      responses:
        '200':
          description: Philosophy file content
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
                  content:
                    type: string
                  path:
                    type: string
                  lastModified:
                    type: string
                    format: date-time
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags:
        - Workspaces
      summary: Update project philosophy file
      parameters:
        - $ref: '#/components/parameters/WorkspaceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - content
              properties:
                content:
                  type: string
                  description: Markdown content for the philosophy file
      responses:
        '200':
          description: Philosophy file updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  path:
                    type: string
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/agents/{agentId}:
    get:
      tags:
        - Agents
      summary: Get agent template details
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent details with content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentDetails'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    put:
      tags:
        - Agents
      summary: Update agent template
      parameters:
        - $ref: '#/components/parameters/AgentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                content:
                  type: string
                  description: Updated markdown content
      responses:
        '200':
          description: Agent updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
    delete:
      tags:
        - Agents
      summary: Delete agent template
      parameters:
        - $ref: '#/components/parameters/AgentId'
      responses:
        '200':
          description: Agent deleted
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /api/search:
    get:
      tags:
        - Search
      summary: Global search
      parameters:
        - name: q
          in: query
          required: true
          description: Search query
          schema:
            type: string
        - name: type
          in: query
          description: Filter by type (plans, workspaces, agents)
          schema:
            type: string
            enum: [all, plans, workspaces, agents]
            default: all
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResults'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  parameters:
    WorkspaceId:
      name: workspaceId
      in: path
      required: true
      description: Workspace ID
      schema:
        type: string
        example: ds_my-project-abc123
...
    AgentId:
      name: agentId
      in: path
      required: true
      description: Agent template ID (filename without .agent.md)
      schema:
        type: string
        example: coordinator
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Resource not found
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: Invalid request parameters
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            type: object
            properties:
              error:
                type: string
                example: An unexpected error occurred
  schemas:
    Workspace:
      type: object
      properties:
        id:
          type: string
          example: ds_my-project-abc123
        name:
          type: string
          example: My Project
        path:
          type: string
          example: /Users/dev/my-project
        registeredAt:
          type: string
          format: date-time
        lastAccessed:
          type: string
          format: date-time
        activePlanCount:
          type: integer
        archivedPlanCount:
          type: integer
        indexed:
          type: boolean
    WorkspaceDetails:
      allOf:
        - $ref: '#/components/schemas/Workspace'
        - type: object
          properties:
            plans:
              type: array
              items:
                $ref: '#/components/schemas/PlanSummary'
            deployedAgents:
              type: array
              items:
                type: string
            profile:
              type: object
              description: Workspace profile/indexing data
    PlanSummary:
      type: object
      properties:
        id:
          type: string
          example: plan_abc123_xyz789
        workspaceId:
          type: string
        title:
          type: string
          example: Add Authentication
        priority:
          type: string
          enum: [low, medium, high, critical]
        status:
          type: string
          enum: [active, paused, completed, archived, failed]
        category:
          type: string
          enum: [feature, bug, change, analysis, debug, refactor, documentation]
        currentAgent:
          type: string
          nullable: true
        stepCount:
          type: integer
        completedStepCount:
          type: integer
        updatedAt:
          type: string
          format: date-time
...
    Agent:
      type: object
      properties:
        id:
          type: string
          example: coordinator
        name:
          type: string
          example: Coordinator
        description:
          type: string
        lastModified:
          type: string
          format: date-time
        deployedCount:
          type: integer
        isTemplate:
          type: boolean
    AgentDetails:
      allOf:
        - $ref: '#/components/schemas/Agent'
        - type: object
          properties:
            content:
              type: string
              description: Raw markdown content
            deployments:
              type: array
              items:
                type: object
                properties:
                  workspaceId:
                    type: string
                  workspaceName:
                    type: string
                  syncStatus:
                    type: string
                    enum: [synced, outdated, customized, missing]
                  hash:
                    type: string
...
    SearchResults:
      type: object
      properties:
        query:
          type: string
        total:
          type: integer
        results:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
                enum: [plan, workspace, agent]
              id:
                type: string
              title:
                type: string
              description:
                type: string
              matches:
                type: array
                items:
                  type: object
                  properties:
                    field:
                      type: string
                    highlight:
                      type: string